KOA-4801 Move story to separate file
